+++ R/api.R
@@ -168,9 +168,10 @@ swecris_fundings <- function() {
 
 stripname <- function(x, name) {
   thisdepth <- depth(x)
+  nameIndex <- which(names(x) == name)
   if (thisdepth == 0) {
     return (x)
-  } else if (length(nameIndex <- which(names(x) == name))) {
+  } else if (length(nameIndex)) {
     x <- x[-nameIndex]
   }
   return(lapply(x, stripname, name))
@@ -194,7 +195,8 @@ parse_swecris_dates <- function(x) {
   x |>
     mutate(across(
       any_of(c("loadedDate", "updatedDate")),
-      \(x) lubridate::parse_date_time(x, "bdY HM")
+      \(x) lubridate::ymd_hms(x, truncated = 2)
+#      \(x) lubridate::parse_date_time(x, "bdY HM")
     )) |>
     mutate(across(
       any_of(c("projectStartDate", "projectEndDate")),
@@ -202,7 +204,7 @@ parse_swecris_dates <- function(x) {
     )) |>
     mutate(across(
       any_of(c("fundingStartDate", "fundingEndDate")),
-      \(x) lubridate::parse_date_time(x, c("Ymd", "Y"))
+      \(x) lubridate::parse_date_time(x, c("Ymd", "Y", "Ymd HMS"))
     )) |>
     mutate(across(
       any_of(c("fundingYear")),