+++ NEWS.md
@@ -1,3 +1,11 @@
+# swecris 0.1.8
+
+* Fix to improve parsing of date formats
+* Refresh of embedded swecris_kth data
+* Deprecation of older fcns for Finnish and Norwegian lists
+* New fcns for retrieving Norwegian lists for publishers and journals
+
+
 # swecris 0.1.7
 
 * Fix to improve parsing of varying date formats